+++ jeap-archrepo-metamodel/src/main/java/ch/admin/bit/jeap/archrepo/metamodel/relation/EventRelation.java
@@ -7,13 +7,14 @@ import jakarta.persistence.Entity;
 import lombok.*;
 
 @EqualsAndHashCode(callSuper = true)
-@ToString(callSuper = true)
+@ToString(callSuper = true, onlyExplicitlyIncluded = true)
 @NoArgsConstructor
 @Getter
 @Entity
 @DiscriminatorValue("EVENT")
 public class EventRelation extends AbstractRelation {
     @NonNull
+    @ToString.Include
     private String eventName;
 
     @Override