+++ jeap-archrepo-metamodel/src/main/java/ch/admin/bit/jeap/archrepo/metamodel/system/SystemGraph.java
@@ -10,7 +10,8 @@ import org.hibernate.annotations.JdbcTypeCode;
 import java.util.UUID;
 
 @EqualsAndHashCode(callSuper = false)
-@ToString
+// Include-only so that the graph blob can never reach a log line - see EntityToStringDoesNotQueryTest
+@ToString(onlyExplicitlyIncluded = true)
 @Entity
 @Getter
 @Table(name = "system_graph")
@@ -18,10 +19,12 @@ public class SystemGraph extends MutableDomainEntity {
 
     @Id
     @NotNull
+    @ToString.Include
     private UUID id;
 
     @NotNull
     @Column(name = "system_name")
+    @ToString.Include
     private String systemName;
 
     @Lob
@@ -29,6 +32,7 @@ public class SystemGraph extends MutableDomainEntity {
     @Column(name = "graph_data", columnDefinition = "BYTEA")
     private byte[] graphData;
 
+    @ToString.Include
     private String fingerprint;
 
     protected SystemGraph() {